Fix sevice card overflow on Firefox mobile and Android WebView
I recently noticed that the service card is overflowing. It is happening
only on Firefox mobile, and on Androind apps that display web pages, I
assume with WebView. I guess that at some point somebody made a change
that somewhere in the HTML rendering engine of some core component and
it resulted in this bug.

I tested the fix on Firefox and Chrome on Windows both on small and
large screen sizes and it works as expected.